pyproject.toml
[project]
name = "duke-ml-chatbot"
version = "0.1.0"
description = "Progressive AI chatbot tutorial - from terminal to web UI"
requires-python = ">=3.11"
dependencies = [
    "openai==2.24.0",
    "python-dotenv==1.1.1",
    "flask==3.1.3",
    "gradio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
